requirements.txt
pandas
numpy
requests
python-dotenv
pymongo
scikit-learn
xgboost
matplotlib
seaborn
shap
fastapi
uvicorn
streamlit
loguru
pydantic
tenacity
evidently
joblib
plotly

